add correct licensing: ## License
This project consists of two parts with different licenses:

### Project Code
The code in this project is licensed under "The Unlicense" - released to the public domain. See the [LICENSE](LICENSE) file for details.

### Pandoc
This project uses [Pandoc](https://pandoc.org/), which is licensed under GPL-2.0-or-later. Pandoc's licensing information:
- [Pandoc Copyright](https://github.com/jgm/pandoc/blob/main/COPYRIGHT)
- [Pandoc License](https://github.com/jgm/pandoc/blob/main/COPYING.md)

By using this project, you agree to comply with the terms of both licenses.